+++ kubernetes-gitlab-sonar-lab/kubernetes-gitlab-sonar-lab/infra/platform/providers.tf
@@ -1,9 +1,0 @@
-provider "kubernetes" {
-  config_path = pathexpand(var.kubeconfig_path)
-}
-
-provider "helm" {
-  kubernetes = {
-    config_path = pathexpand(var.kubeconfig_path)
-  }
-}